enabled reuse of packaging packets for messages
