+++ Sources/TDS/Connection/TDSConnection+RewrapTLS.swift
@@ -42,7 +42,7 @@ public final class PipelineOrganizationHandler: ChannelDuplexHandler, RemovableC
             
             switch message.headerType {
             case .prelogin:
-                let message = try ByteBuffer(unpackingSSLPayloadFrom: message, allocator: context.channel.allocator)
+                let message = try ByteBuffer(unpackingDataFrom: message, allocator: context.channel.allocator)
                 sslHandshakeState.addReceivedData(message)
                 self.state = .sslHandshake(sslHandshakeState)
                 context.fireChannelRead(self.wrapInboundOut(sslHandshakeState.inputBuffer))
@@ -72,7 +72,7 @@ public final class PipelineOrganizationHandler: ChannelDuplexHandler, RemovableC
     private func _flush(context: ChannelHandlerContext) throws {
         switch self.state {
         case .sslHandshake(var sslHandshakeState):
-            let message = try TDSMessage(packingSSLPayloadWith: &sslHandshakeState.outputBuffer, allocator: context.channel.allocator)
+            let message = try TDSMessage(packingDataWith: &sslHandshakeState.outputBuffer, headerType: .prelogin, allocator: context.channel.allocator)
             context.writeAndFlush(self.wrapOutboundOut(message), promise: sslHandshakeState.outputPromise)
             sslHandshakeState.outputBuffer.clear()
             state = .sslHandshake(sslHandshakeState)
